pyproject.toml
[project]
name = "fastapi-til"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "py-ulid (>=1.0.3,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
